gitminder: switch to 'locate'
'locate' is much faster than 'find' and performs better for large
filesystems. Unfortunately, it can't discern between file types, so we
need to manually test for dir-ness. Then we can use grep for the rest of
the filtering.

On my disk with 340GB used, the previous 'find'-based version took 11
seconds to run, and the 'locate'-based version after this commit takes 3
seconds to run.

This commit changes the frontend semantics of 'gitminder'.

TODO: just take in --include-regex and --exclude-regex instead of this
wonky half and half behavior.